refactor(detect): split 2,219-line pipeline.rs into stage modules
pipeline.rs concentrated parsing, framework detection, partitioning,
reduction, Wolfi resolution, and source scanning in one file. Split the
stage internals into focused submodules under src/pipeline/, leaving the
parent module with just the public API, the orchestrator that sequences the
stages, and the test suite:

- pipeline/tree.rs        - filesystem walk + file classification
- pipeline/framework.rs   - framework detection pass
- pipeline/partitioning.rs- service-bucket partitioning + workspace expansion
- pipeline/reducing.rs    - bucket -> UniversalBuild reduction
- pipeline/wolfi.rs       - Wolfi package version resolution
- pipeline/scan.rs        - version-file scanning

Pure code move: no behavior change. All 291 detect tests pass; clippy and
fmt clean.
